+++ src/main/java/org/oscm/identity/controller/MainController.java
@@ -12,9 +12,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.oscm.identity.error.IdentityProviderException;
+import org.oscm.identity.model.request.TokenValidationRequest;
 import org.oscm.identity.oidc.request.*;
-import org.oscm.identity.oidc.response.validation.AuthTokenValidator;
-import org.oscm.identity.oidc.response.validation.TokenValidationResult;
+import org.oscm.identity.model.request.validation.AuthTokenValidator;
+import org.oscm.identity.model.request.validation.TokenValidationResult;
 import org.oscm.identity.oidc.tenant.TenantConfiguration;
 import org.oscm.identity.service.TenantService;
 import org.springframework.beans.factory.annotation.Autowired;
