+++ src/main/java/org/oscm/identity/controller/UserController.java
@@ -13,9 +13,9 @@ import org.json.JSONObject;
 import org.oscm.identity.oidc.request.DefaultGetUserRequest;
 import org.oscm.identity.oidc.request.RequestHandler;
 import org.oscm.identity.oidc.request.UserRequest;
-import org.oscm.identity.oidc.response.ResponseHandler;
-import org.oscm.identity.oidc.response.UserInfoResponse;
-import org.oscm.identity.oidc.response.mapper.ResponseMapper;
+import org.oscm.identity.model.response.ResponseHandler;
+import org.oscm.identity.model.response.UserInfoResponse;
+import org.oscm.identity.model.response.mapper.ResponseMapper;
 import org.oscm.identity.oidc.tenant.TenantConfiguration;
 import org.oscm.identity.service.TenantService;
 import org.springframework.beans.factory.annotation.Autowired;
