+++ src/main/java/org/oscm/identity/model/request/validation/NonceValidationStrategy.java
@@ -7,10 +7,10 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.identity.oidc.response.validation;
+package org.oscm.identity.model.request.validation;
 
 import lombok.extern.slf4j.Slf4j;
-import org.oscm.identity.oidc.request.TokenValidationRequest;
+import org.oscm.identity.model.request.TokenValidationRequest;
 import org.springframework.stereotype.Component;
 
 import javax.xml.bind.ValidationException;
