+++ src/main/java/org/oscm/identity/model/request/validation/TokenValidationStrategy.java
@@ -7,9 +7,9 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.identity.oidc.response.validation;
+package org.oscm.identity.model.request.validation;
 
-import org.oscm.identity.oidc.request.TokenValidationRequest;
+import org.oscm.identity.model.request.TokenValidationRequest;
 
 import javax.xml.bind.ValidationException;
