+++ src/main/java/org/oscm/identity/model/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package org.oscm.identity.oidc.response;
+package org.oscm.identity.model.response;
 
 import lombok.Data;
