+++ src/main/java/org/oscm/identity/model/response/UserGroupResponse.java
@@ -1,0 +1,11 @@
+package org.oscm.identity.model.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder(builderMethodName = "of")
+public class UserGroupResponse {
+
+    private String name;
+}
