+++ src/test/java/org/oscm/identity/controller/MainControllerTest.java
@@ -19,8 +19,8 @@ import org.oscm.identity.oidc.request.DefaultRequestManager;
 import org.oscm.identity.oidc.request.RequestHandler;
 import org.oscm.identity.oidc.request.RequestManager;
 import org.oscm.identity.oidc.request.TokenRequest;
-import org.oscm.identity.oidc.response.validation.AuthTokenValidator;
-import org.oscm.identity.oidc.response.validation.TokenValidationResult;
+import org.oscm.identity.model.request.validation.AuthTokenValidator;
+import org.oscm.identity.model.request.validation.TokenValidationResult;
 import org.oscm.identity.oidc.tenant.TenantConfiguration;
 import org.oscm.identity.service.TenantService;
 import org.springframework.http.HttpStatus;
