+++ src/test/java/org/oscm/identity/oidc/response/validation/NonceValidationStrategyTest.java
@@ -13,7 +13,9 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.oscm.identity.oidc.request.TokenValidationRequest;
+import org.oscm.identity.model.request.TokenValidationRequest;
+import org.oscm.identity.model.request.validation.NonceValidationStrategy;
+import org.oscm.identity.model.request.validation.TokenValidationStrategy;
 
 import javax.xml.bind.ValidationException;
